+++ src/app/containers/IndexPageSection/utilities/filterAllowedItems.jsx
@@ -2,7 +2,7 @@ import dropWhile from 'ramda/src/dropWhile';
 import {
   getAssetTypeCode,
   getHeadlineUrlAndLive,
-} from '#lib/utilities/getStoryPromoInfo';
+} from '#lib/utilities/getUrlHelpers';
 
 const MAX_ALLOWED_ITEMS_FIRST_SECTION = 13;
 const MAX_ALLOWED_ITEMS = 10;
